Change Tx broadcaster base to WalletActionViewModel
