+++ WalletWasabi.Gui/Controls/WalletExplorer/TransactionBroadcasterView.xaml
@@ -1,0 +1,11 @@
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:controls="clr-namespace:WalletWasabi.Gui.Controls;assembly=WalletWasabi.Gui"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450">
+  <controls:GroupBox Title="{Binding Title}" TextBlock.FontSize="30" Padding="20" Margin="10">
+    <Grid Classes="content">
+    </Grid>
+  </controls:GroupBox>
+</UserControl>
