--- WalletWasabi.Gui/Tabs/TransactionBroadcasterView.xaml.cs
@@ -1,19 +1,0 @@
-﻿using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace WalletWasabi.Gui.Tabs
-{
-	public class TransactionBroadcasterView : UserControl
-	{
-		public TransactionBroadcasterView()
-		{
-			InitializeComponent();
-		}
-
-		private void InitializeComponent()
-		{
-			AvaloniaXamlLoader.Load(this);
-		}
-	}
-}
